+++ Examples/FSharp/GiftCard/RedeemBinanceCode_Example.fs
@@ -1,0 +1,24 @@
+open System
+open System.Net
+open System.Net.Http
+open System.Threading.Tasks
+open Microsoft.Extensions.Logging
+open Binance.Common
+open Binance.Spot
+open Binance.Spot.Models
+
+[<EntryPoint>]
+let main argv =
+    let loggerFactory = LoggerFactory.Create(fun (builder:ILoggingBuilder) -> 
+        builder.AddConsole() |> ignore
+    )
+    let logger = loggerFactory.CreateLogger()
+
+    let loggingHandler = new BinanceLoggingHandler(logger)
+    let httpClient = new HttpClient(loggingHandler)
+    
+    let giftCard = new GiftCard(httpClient)
+    
+    let result = giftCard.RedeemBinanceCode("X1X1X1X1X1X11XX1X11X1") |> Async.AwaitTask |> Async.RunSynchronously
+    
+    0